Use entire patch for aligning the code affinities
Overrides the hard-coded central crop of 20% of the pixels.
To avoid memory issues, the patch size is reduced from 100*100 to
20*20.
To compensate for the fewer pixels in each patch, the batch size
has been increased from 10 to 25 (50 crashes due to the memory
issues associated with creating the batch_size * c_patch_size**2 *
c_patch_size**2 *c_patch_size**2 matrix) and also increase the
number of batches in each epoch from 10 to 100.

TODO: Update code so that the CONFIG parameter
"affinity_patch_size" is actually used for these calculations.