Start moving hard-coded parameters and assumptions into CONFIG
Enables setting parameters to control if entire patch or a 20 %
centre crop should be used for code alignment. The previous hard-
coded assumption is designed to work with the case where patches
are 100*100 pixels. Now the entire patch will be used if the patch
size is equal to the affinity patch size, if not, the crop percent
will be calculated based on these two parameters.

Also start basing the difference image on the original data rather
than the reconstructed data (the difference is with the transformed
data from the other domain).

Also test using a global kernel size for affinities.
TODO: Estimate this.