Create additional change map with different params and save CONFIG
The additional change map is implemented in a slightly "hacky"
manner, but is intended to be further developed to study the
effect of the threshold for the final confusion map.

Also start work towards aligning how the images are saved.

CONFIG is now saved as a text file to preserve processing
parameters.